feat: 抽离 action 公共逻辑
