+++ src/onebot11/actions/GetGroupInfo.ts
@@ -1,8 +1,8 @@
 import { OB11Response } from "./utils";
-import { BaseCheckResult } from "./types";
 import { OB11Group, OB11Return } from '../types';
 import { getGroup, groups } from "../../common/data";
 import { OB11Constructor } from "../constructor";
+import BaseAction from "./BaseAction";
 
 export type ActionType = 'get_group_info'
 
@@ -13,23 +13,8 @@ export interface PayloadType {
 
 export type ReturnDataType = OB11Group[]
 
-class GetGroupInfo {
+class GetGroupInfo extends BaseAction {
     static ACTION_TYPE: ActionType = 'get_group_info'
-
-    async check(jsonData: any): Promise<BaseCheckResult> {
-        return {
-            valid: true,
-        }
-    }
-
-    async handle(jsonData: any) {
-        const result = await this.check(jsonData)
-        if (!result.valid) {
-            return OB11Response.error(result.message)
-        }
-        const resData = await this._handle(jsonData)
-        return resData
-    }
 
     async _handle(payload: PayloadType): Promise<OB11Return<ReturnDataType | null>> {
         const group = await getGroup(payload.group_id.toString())
