+++ src/onebot11/actions/GetGroupMemberInfo.ts
@@ -1,8 +1,8 @@
 import { OB11Response } from "./utils";
-import { BaseCheckResult } from "./types";
 import { OB11GroupMember, OB11Return } from '../types';
 import { getGroupMember } from "../../common/data";
 import { OB11Constructor } from "../constructor";
+import BaseAction from "./BaseAction";
 
 export type ActionType = 'get_group_member_info'
 
@@ -14,23 +14,8 @@ export interface PayloadType {
 
 export type ReturnDataType = OB11GroupMember
 
-class GetGroupMemberInfo {
+class GetGroupMemberInfo extends BaseAction {
     static ACTION_TYPE: ActionType = 'get_group_member_info'
-
-    async check(jsonData: any): Promise<BaseCheckResult> {
-        return {
-            valid: true,
-        }
-    }
-
-    async handle(jsonData: any) {
-        const result = await this.check(jsonData)
-        if (!result.valid) {
-            return OB11Response.error(result.message)
-        }
-        const resData = await this._handle(jsonData)
-        return resData
-    }
 
     async _handle(payload: PayloadType): Promise<OB11Return<ReturnDataType | null>> {
         const member = await getGroupMember(payload.group_id.toString(), payload.user_id.toString())
