+++ src/onebot11/actions/GetGroupMemberList.ts
@@ -1,9 +1,9 @@
 import { OB11Response } from "./utils";
-import { BaseCheckResult } from "./types";
 import { OB11GroupMember, OB11Return } from '../types';
 import { getGroup } from "../../common/data";
 import { NTQQApi } from "../../ntqqapi/ntcall";
 import { OB11Constructor } from "../constructor";
+import BaseAction from "./BaseAction";
 
 export type ActionType = 'get_group_member_list'
 
@@ -14,23 +14,8 @@ export interface PayloadType {
 
 export type ReturnDataType = OB11GroupMember[]
 
-class GetGroupMemberList {
+class GetGroupMemberList extends BaseAction {
     static ACTION_TYPE: ActionType = 'get_group_member_list'
-
-    async check(jsonData: any): Promise<BaseCheckResult> {
-        return {
-            valid: true,
-        }
-    }
-
-    async handle(jsonData: any) {
-        const result = await this.check(jsonData)
-        if (!result.valid) {
-            return OB11Response.error(result.message)
-        }
-        const resData = await this._handle(jsonData)
-        return resData
-    }
 
     async _handle(payload: PayloadType): Promise<OB11Return<ReturnDataType | null>> {
         const group = await getGroup(payload.group_id.toString());
